LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/rowGap-expected.txt


PASS row-gap responsive to style changes
PASS row-gap clamped to 0px on keyframes
FAIL row-gap responsive to inherited changes assert_equals: expected "80px" but got "normal"
